Fixed bug with new coming users

+++ src/main/kotlin/admin_bot_functions/statistic/chains/PointStartDateChainFinal.kt
@@ -50,7 +50,7 @@ class PointStartDateChainFinal(
                                 )
                             ).convertToVertical()
                         ),
-                        mEditingMessageId = mStates.state(updating).long("statMessageId")
+                        mEditingMessageId = mStates.state(updating).int("statMessageId").toLong()
                     )
                 )
             }
